+++ src/component/ModuleContainer.ts
@@ -1,0 +1,5 @@
+import styled from "@alethio/ui/lib/styled-components";
+
+export const ModuleContainer = styled.div`
+    position: relative;
+`;