+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# Unreleased
+
+- Better plugin error handling and add config validation
+
 # v1.0.0-beta.5
 
 - Add support for inline (baked-in) plugins