+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Better plugin error handling and add config validation
 - Add alias support for data adapter config (ref variant)
+- **Breaking change**: Modules are now wrapped with a `position: relative` container. Any module that had elements depending on the parent layout, such as flex or `height: 100%` could theoretically break. As a workaround, a new method called `getWrapperStyle` was added to the module definition.
 
 # v1.0.0-beta.5