+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Unreleased
+# v1.0.0-beta.6
 
 - Better plugin error handling and add config validation
 - Add alias support for data adapter config (ref variant)